deathmatch - bomb delay increased

+++ ogre/src/cheats/CheatBomb.h
@@ -41,7 +41,7 @@ private:
     bool mIsBombInProgress;
     bool mIsBombJumpsInProgress;
     bool mIsBombExplosionInProgress;
-    Ogre::uint8 mBlowCounter;
+    Ogre::uint16 mBlowCounter;
     Ogre::uint8 mExplosionCounter;
     PhysicsVehicle * mPlayerVehicle;
